# OpenAPI.yaml
openapi: 3.0.3
info:
  title: HR Portal API
  version: 1.0.0
servers:
  - url: http://localhost:8000
paths:
...
  /departments:
    get:
      summary: List departments
      operationId: listDepartments
      responses:
        '200':
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Department'
    post:
      summary: Create department
      operationId: createDepartment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentCreate'
      responses:
        '201':
          description: Department created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
  /departments/{dept_no}:
    get:
      summary: Get department by ID
      operationId: getDepartment
      parameters:
        - in: path
          name: dept_no
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Department
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        '404':
          description: Department not found
    put:
      summary: Update department
      operationId: updateDepartment
      parameters:
        - in: path
          name: dept_no
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentUpdate'
      responses:
        '200':
          description: Updated department
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        '404':
          description: Department not found
    delete:
      summary: Delete department
      operationId: deleteDepartment
      parameters:
        - in: path
          name: dept_no
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Department deleted
        '404':
          description: Department not found
...
components:
  schemas:
...
    Department:
      type: object
      required: [dept_no, dept_name]
      properties:
        dept_no:
          type: string
          maxLength: 4
        dept_name:
          type: string
          maxLength: 40
    DepartmentCreate:
      type: object
      required: [dept_no, dept_name]
      properties:
        dept_no:
          type: string
        dept_name:
          type: string
    DepartmentUpdate:
      type: object
      properties:
        dept_name:
          type: string
...